move the logic fro conftest.py to tools.util
...and keep only the pytest.fixture definition
